Remove file uploading from CadViewer to src/utils/loader.
