+++ src/Containers/CadView.jsx
@@ -22,6 +22,7 @@ import useStore from '../store/useStore'
 import {computeElementPathIds, setupLookupAndParentLinks} from '../utils/TreeUtils'
 import {assertDefined} from '../utils/assert'
 import {handleBeforeUnload} from '../utils/event'
+import {loadLocalFile, getUploadedBlobPath} from '../utils/loader'
 import {getDownloadURL, parseGitHubRepositoryURL} from '../utils/GitHub'
 import SearchIndex from './SearchIndex'
 import {usePlaceMark} from '../hooks/usePlaceMark'
@@ -261,7 +262,7 @@ export default function CadView({
     const uploadedFile = pathPrefix.endsWith('new')
 
     if (uploadedFile) {
-      filepath = getNewModelRealPath(filepath)
+      filepath = getUploadedBlobPath(filepath)
       debug().log('CadView#loadIfc: parsed blob: ', filepath)
       window.addEventListener('beforeunload', handleBeforeUnload)
     }
@@ -313,30 +314,6 @@ export default function CadView({
 
     debug().error('CadView#loadIfc: Model load failed!')
     return loadedModel
-  }
-
-
-  /** Upload a local IFC file for display. */
-  function loadLocalFile() {
-    const viewerContainer = document.getElementById('viewer-container')
-    const fileInput = document.createElement('input')
-    fileInput.setAttribute('type', 'file')
-    fileInput.addEventListener(
-        'change',
-        (event) => {
-          debug().log('CadView#loadLocalFile#event:', event)
-          let ifcUrl = URL.createObjectURL(event.target.files[0])
-          debug().log('CadView#loadLocalFile#event: ifcUrl: ', ifcUrl)
-          const parts = ifcUrl.split('/')
-          ifcUrl = parts[parts.length - 1]
-          window.removeEventListener('beforeunload', handleBeforeUnload)
-          navigate(`${appPrefix}/v/new/${ifcUrl}.ifc`)
-        },
-        false,
-    )
-    viewerContainer.appendChild(fileInput)
-    fileInput.click()
-    viewerContainer.removeChild(fileInput)
   }
 
 
@@ -619,10 +596,9 @@ export default function CadView({
           },
         }}
         >
-          {isSearchBarVisible &&
-            <SearchBar
-              fileOpen={loadLocalFile}
-            />}
+          {
+            isSearchBarVisible &&
+              <SearchBar fileOpen={() => loadLocalFile(navigate, appPrefix, handleBeforeUnload)}/>}
           {
             modelPath.repo !== undefined &&
             <BranchesControl location={location}/>
@@ -758,18 +734,4 @@ const getFinalURL = async (url, accessToken) => {
     default:
       return url
   }
-}
-
-
-/**
- * @param {string} filepath
- * @return {string}
- */
-export function getNewModelRealPath(filepath) {
-  const l = window.location
-  filepath = filepath.split('.ifc')[0]
-  const parts = filepath.split('/')
-  filepath = parts[parts.length - 1]
-  filepath = `blob:${l.protocol}//${l.hostname + (l.port ? `:${l.port}` : '')}/${filepath}`
-  return filepath
 }
